backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv==1.0.0

anthropic==0.18.1
openai==1.12.0

# For OAuth
requests==2.31.0
